+++ sabka-bazar/src/components/cart/cart.scss
@@ -1,6 +1,6 @@
 .shopping-cart {
    width: 100%;
-   max-height: 80%;
+   max-height: 100%;
     float: right;
     background-color: white;
     border-radius: 3px;
@@ -178,7 +178,10 @@
 
 @media screen and (min-width: 480px){
     .shopping-cart{
-        width: 40%;
+        position: absolute;
+        width: 30%;
+        top:15%;
+        left:60%;
     }
     
 }
