+++ sabka-bazar/src/components/model/index.js
@@ -1,14 +1,14 @@
 import React from "react";
 import "./model.scss";
-import Cart from "../cart";
+import Cart from "../../containers/cart";
 
 function Model(props) {
-  const { show, handleClose, cartProducts, plusProduct, minusProduct } = props;
+  const { show, handleClose } = props;
 
   const showHideClassName = show ? "modal display-block" : "modal display-none";
 
   return <div className={`modal-main ${showHideClassName}`}>
-    <Cart handleClose={handleClose} cartProducts={cartProducts} plusProduct={plusProduct} minusProduct={minusProduct} />
+    <Cart handleClose={handleClose} />
   </div>;
 }
