+++ sabka-bazar/src/components/model/model.scss
@@ -1,17 +1,17 @@
 .modal {
     position: fixed;
-    top: 0;
-    left: 0;
+    //top: 50%;
+    //left: 50%;
     width:100%;
     height: 100%;
     background: rgba(0, 0, 0, 0.6);
   }
   
   .modal-main {
-    position:fixed;
-    background:transparent;
-    width: 80%;
-    height: auto;
+    //position:fixed;
+    //background:transparent;
+    //width: 100%;
+    //height: 100%;
     top:50%;
     left:50%;
     transform: translate(-50%,-50%);
