+++ sabka-bazar/src/containers/login/index.js
@@ -1,9 +1,9 @@
 import Login from '../../components/login';
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import { requestAddProduct, requestMinusProduct } from "../plp/actions";
+// import { requestAddProduct, requestMinusProduct } from "../plp/actions";
 
 const mapStateToProps = state => ({ cartProducts: state.plp.cartProducts });
-const mapDispatchToProps = dispatch => bindActionCreators({ requestAddProduct, requestMinusProduct }, dispatch);
+//const mapDispatchToProps = dispatch => bindActionCreators({ requestAddProduct, requestMinusProduct }, dispatch);
 
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+export default connect(mapStateToProps, {})(Login);
