+++ sabka-bazar/src/containers/plp/index.js
@@ -1,13 +1,13 @@
 import Plp from '../../components/plp';
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import { requestProductsData, requestaddToCart, requestAddProduct, requestMinusProduct } from "./actions";
+import { requestProductsData, requestaddToCart } from "./actions";
 
 const mapStateToProps = (state) => ({
     products: state.plp.products,
     categories: state.home.categories,
     cartProducts: state.plp.cartProducts
 });
-const mapDispatchToProps = (dispatch) => bindActionCreators({ requestProductsData, requestaddToCart, requestAddProduct, requestMinusProduct }, dispatch);
+const mapDispatchToProps = (dispatch) => bindActionCreators({ requestProductsData, requestaddToCart }, dispatch);
 
 export default connect(mapStateToProps, mapDispatchToProps)(Plp);
