+++ sabka-bazar/src/containers/register/index.js
@@ -1,9 +1,9 @@
 import Register from '../../components/register';
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import { requestAddProduct, requestMinusProduct } from "../plp/actions";
+// import { requestAddProduct, requestMinusProduct } from "../plp/actions";
 
 const mapStateToProps = state => ({ cartProducts: state.plp.cartProducts });
-const mapDispatchToProps = dispatch => bindActionCreators({ requestAddProduct, requestMinusProduct }, dispatch);
+// const mapDispatchToProps = dispatch => bindActionCreators({ requestAddProduct, requestMinusProduct }, dispatch);
 
-export default connect(mapStateToProps, mapDispatchToProps)(Register);
+export default connect(mapStateToProps, {})(Register);
